revert: dashboard IS settings, portfolio IS the trade-history page
The previous restructure split things up (and then tried to merge
them) — wrong both times. The user's model is simpler:

  Dashboard (/account)            = settings/editors page
                                    (mandate, agents, visibility,
                                    buy mandate, sell mandate)
  Portfolio (/portfolios/<slug>)  = the portfolio with holdings + trades

Two pages, one purpose each. Nav: Dashboard + Portfolio sit side by
side and go to genuinely different places.

Changes:

* `/account/page.tsx` — restored to the editor-cards view it originally
  was, with the sell-mandate card added (the only intentional addition
  from the recent dance). Header shows the portfolio name and a
  "View portfolio →" link.

* `/account/settings/page.tsx` — DELETED. Dashboard IS settings, so a
  separate subpage was redundant.

* `/portfolios/[slug]/page.tsx` — removed the "Settings →" chip from
  the owner-only header row. Owners reach settings via the Dashboard
  nav link.

The proxy.ts is already restored (signed-in `/` → marketing). Nav is
already restored (Dashboard | Watchlist | Portfolio | ...). This
commit just finishes the cleanup by killing the settings subpage and
its stale link.
